terraform/cloud_functions/.staging/beer-recommender/requirements.txt
google-cloud-bigquery==3.17.2
PyJWT==2.8.0
